app/requirements.txt
celery==5.4.0
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1

psycopg2-binary==2.9.6
redis==5.0.7
drf-yasg==1.21.7
